requirements.txt
limacharlie>=4.11.0
PyYAML>=6.0
requests>=2.28.0
stix2>=3.0.0
jinja2>=3.1.0
